Hardcode which files should have what null padding
It appears that row lengths should be a multiple of four in a short list
of files, but be two more than a multiple of four in most other files. I
haven't found any indication in the files themselves of what the
alignment of these should look like, so I am just hardcoding a list of
names for now.